improve entity
